refactor: sentence transformer module + update docs
