--- docs/rag/retrieval/text_retrieval/nv_embed_2.md
@@ -1,3 +1,0 @@
-# NV-Embed-v2 Retrieval
-
-::: medrag_multi_modal.retrieval.text_retrieval.nv_embed_2
