+++ docs/rag/retrieval/text_retrieval/sentence_transformer.md
@@ -1,0 +1,3 @@
+# Sentence Transformer Retrieval
+
+::: medrag_multi_modal.retrieval.text_retrieval.sentence_transformer_retrieval
