+++ medrag_multi_modal/retrieval/text_retrieval/__init__.py
@@ -1,7 +1,10 @@
 from .bm25s_retrieval import BM25sRetriever
 from .contriever_retrieval import ContrieverRetriever
 from .medcpt_retrieval import MedCPTRetriever
-from .nv_embed_2 import NVEmbed2Retriever, SentenceTransformerRetriever
+from .sentence_transformer_retrieval import (
+    NVEmbed2Retriever,
+    SentenceTransformerRetriever,
+)
 
 __all__ = [
     "BM25sRetriever",
