+++ medrag_multi_modal/retrieval/text_retrieval/sentence_transformer_retrieval.py
